% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_pt_biserial_cor.R
\name{abs_pt_biserial_cor}
\alias{abs_pt_biserial_cor}
\title{Title}
\usage{
abs_pt_biserial_cor(w, binary_cov)
}
\arguments{
\item{w}{TBD}

\item{binary_cov}{TBD}
}
\value{
value
}
\description{
Title
}
