% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_cov_bal.R
\name{all_cov_bal}
\alias{all_cov_bal}
\title{Title}
\usage{
all_cov_bal(pseudo_pop, w, c_unordered_vars, ci_appr, all_cov_names, title)
}
\arguments{
\item{pseudo_pop}{TBD}

\item{w}{TBD}

\item{c_unordered_vars}{TBD}

\item{ci_appr}{TBD}

\item{all_cov_names}{TBD}

\item{title}{TBD}
}
\value{
value
}
\description{
Title
}
