% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_cov_bal_boxplot.R
\name{cat_cov_bal_boxplot}
\alias{cat_cov_bal_boxplot}
\title{Title}
\usage{
cat_cov_bal_boxplot(pseudo_pop, ci_appr, var_names, title)
}
\arguments{
\item{pseudo_pop}{TBD}

\item{ci_appr}{TBD}

\item{var_names}{TBD}

\item{title}{TBD}
}
\value{
value
}
\description{
Title
}
