% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_unordered_var.R
\name{cor_unordered_var}
\alias{cor_unordered_var}
\title{Title}
\usage{
cor_unordered_var(w, unordered_var)
}
\arguments{
\item{w}{TBD}

\item{unordered_var}{TBD}
}
\value{
value
}
\description{
Title
}
