% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_subproject.R
\name{finalize_subproject}
\alias{finalize_subproject}
\title{Finalize the processing of the sub project}
\usage{
finalize_subproject(path_obj)
}
\arguments{
\item{path_obj}{TBD}
}
\value{
value
}
\description{
Finalize the processing of the sub project
}
