% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_sub_project.R
\name{initialize_sub_project}
\alias{initialize_sub_project}
\title{Initialize a subproject}
\usage{
initialize_sub_project(
  sp_name,
  pkg = ".",
  external_private = TRUE,
  external_public = TRUE
)
}
\arguments{
\item{sp_name}{}
}
\description{
Initialize a subproject
}
