% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_cov_bal.R
\name{quant_cov_bal}
\alias{quant_cov_bal}
\title{Title}
\usage{
quant_cov_bal(pseudo_pop, ci_appr, var_names, title)
}
\arguments{
\item{pseudo_pop}{TBD}

\item{ci_appr}{TBD}

\item{var_names}{TBD}

\item{title}{TBD}
}
\value{
value
}
\description{
Title
}
