% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_path.R
\name{setup_path}
\alias{setup_path}
\title{Set up required internal and external paths}
\usage{
setup_path(pkg = ".")
}
\arguments{
\item{pkg}{A package name or path.}
}
\value{
value
}
\description{
Sets up required internal and external paths. This include any user defined
external data path, package path and many more.
}
