% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_pseudo_weights.R
\name{summarize_pseudo_weights}
\alias{summarize_pseudo_weights}
\title{Title}
\usage{
summarize_pseudo_weights(pseudo_pop)
}
\arguments{
\item{pseudo_pop}{TBD}
}
\value{
value
}
\description{
Title
}
