% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap_weights.R
\name{cap_weights}
\alias{cap_weights}
\title{Title}
\usage{
cap_weights(
  pseudo_pop,
  ci_appr,
  nthread,
  quant_var_names,
  cat_var_names,
  title
)
}
\arguments{
\item{pseudo_pop}{TBD}

\item{ci_appr}{TBD}

\item{nthread}{TBD}

\item{quant_var_names}{TBD}

\item{cat_var_names}{TBD}

\item{title}{TBD}
}
\value{
value
}
\description{
Title
}
